+++ CHANGES.md
@@ -19,7 +19,7 @@
 
 #### Yocaml_runtime
 
-- Serve .xml, .rss, .atom, .feed files as application/xml
+- Serve .xml, .rss, .atom, .feed files as application/xml (by [reynir](https://reyn.ir/))
 
 ### v2.7.0 2025-11-18 Nantes (France)